+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Set environment variables in a seed (seed toml files only for now)
 
+### Changed
+- Make desktop entry category a configurable option
+
 ## [0.2.0] - 2025-09-27
 
 ### Added